requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.14.0
SQLAlchemy==2.0.35
gunicorn==23.0.0
Werkzeug==3.0.4
Jinja2==3.1.4
itsdangerous==2.2.0
MarkupSafe==3.0.2
click==8.1.7
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==10.3.0